[iOS] Feat: Track row를 3가지 타입으로 구분 후 구현 #210
- VIBE 앱 전체에서 쓰이는 3가지 종류의 row type 구분 후 구현
